Split off Executable into a module.
