+++ lib/chromium-phases/chromium-phases.js
@@ -1,39 +1,11 @@
 // This is where all phases, related to Chromium checkout and build should live.
 // Interfacing with ninja, gclient, git cl? This is gthe spot.
 
-var Promise = require('bluebird');
-var register = require('../core/phase-register.js');
-var types = require('../core/types');
+var register = require('../../core/phase-register.js');
+var types = require('../../core/types');
+var Executable = require('./executable');
+var options = require('../../core/options');
 
-var spawn = require('child_process').spawn;
-var options = require('../core/options');
-
-var chromiumPath = options.chromium;
-if (!chromiumPath)
-  throw new Error('Path to Chromium repo (--chromium=[path/to/cr/src]) is required.');
-
-function Executable(cmd, args) {
-  this.cmd = cmd;
-  this.args = args;
-}
-
-Executable.prototype.toString = function() {
-  return this.cmd + ' ' + this.args.join(' ');
-}
-
-Executable.prototype.run = function() {
-  var child = spawn(this.cmd, this.args, {
-    cwd: chromiumPath,
-    stdio: 'ignore',
-  });
-  child.on('error', function(err) {
-    console.log('Unable to execute "' + this.toString() + '"');
-    console.log('Is your path correct?');
-  });
-  return new Promise(function(resolve) {
-    child.on('close', resolve);
-  })
-}
 
 register({name: 'buildChromium', input: types.string, output: types.string, arity: '1:1', async: true},
   function(data, tags, cb) {
